change Packets to a new script
